Creacion de clases de lectura de fichero y analisis de entrada

+++ src/main/java/clasesProyecto/Onda.java
@@ -1,36 +1,38 @@
 package clasesProyecto;
 
+import enums.Tipos_Ondas;
+
 public class Onda {
 	//duracion  -> calcualdo
-	private float fin;
-	private float inicio;
+	private int fin;
+	private int inicio;
 	private float pico;
 	private Tipos_Ondas tipo;
 	
-	public Onda(float fin, float inicio, float pico, Tipos_Ondas tipo) {
+	public Onda(int fin, int inicio, float pico, Tipos_Ondas tipo) {
 		this.fin = fin;
 		this.inicio = inicio;
 		this.pico = pico;
 		this.tipo = tipo;
 	}
 	
-	public float getDuracion() {
+	public int getDuracion() {
 		return fin - inicio;
 	}
 
-	public float getFin() {
+	public int getFin() {
 		return fin;
 	}
 
-	public void setFin(float fin) {
+	public void setFin(int fin) {
 		this.fin = fin;
 	}
 
-	public float getInicio() {
+	public int getInicio() {
 		return inicio;
 	}
 
-	public void setInicio(float inicio) {
+	public void setInicio(int inicio) {
 		this.inicio = inicio;
 	}
